requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
click==8.1.7
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.1.3
django-cache-url==3.4.5
djangorestframework==3.15.2
drf-spectacular==0.27.2
environs==11.0.0
factory_boy==3.3.1
Faker==30.8.2
inflection==0.5.1
ipdb
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
marshmallow==3.23.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
